+++ pulser/__init__.py
@@ -1,0 +1,7 @@
+"""A pulse-level composer for Pasqal's quantum devices."""
+
+from .register import Register
+
+from .pulses import Pulse
+
+from .sequences import Sequence
