+++ pulser/seq_drawer.py
@@ -1,7 +1,7 @@
 import matplotlib.pyplot as plt
 import numpy as np
 
-from waveforms import ConstantWaveform
+from .waveforms import ConstantWaveform
 
 
 def gather_data(seq):
