+++ pulser/waveforms.py
@@ -1,7 +1,7 @@
 import numpy as np
 import matplotlib.pyplot as plt
 from abc import ABC, abstractmethod
-from utils import validate_duration
+from .utils import validate_duration
 
 
 class Waveform(ABC):
